requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
aiofiles  
hypercorn
locust
databricks-sdk
pytest
pytest-asyncio
logging
statistics